Track rx-timestamp from liburx
Knowning when a valid frame arrived is often important (see get_remote
for an example), and the closer we get the timestamp to Rx, the better.

This adds support for tracking Rx-time the same way the recipe values
are being updated, by registering a pointer to a value which will be
updated upon every new incoming (valid) frame. This way, whenever
handler::recv() succeeds, the timestamp will be updated with a valid
local timestamp.

If it is not being tracked, use timestamp 0 and ignore the field. Tests
and mock are updated as well.