header: move payload field out of struct and provide helper to get offset
The "C approach" is to embed a zero-length array into a struct for the
payload. This is forbidden in C++ so we cannot do this. Instead, only
keep the fields for the actual header in the struct and provide
_get_payload() helpers for each type.

This created some fallout, both logic, size-computation and tests needed
to be updated.